structure: move test factories to test/fixtures/, delete lib/
