+++ nuxt.config.ts
@@ -9,9 +9,9 @@ const relaxedIndexAccess = () => ({
   compilerOptions: { noUncheckedIndexedAccess: false },
 });
 
-// `lib/` and `test/` sit outside every project Nuxt generates, so their
-// files are only checked where an app or server file happens to import them.
-const rootDirsOutsideNuxtProjects = ["../lib/**/*", "../test/**/*"];
+// `test/` sits outside every project Nuxt generates, so its files are only
+// checked where an app or server file happens to import them.
+const rootDirsOutsideNuxtProjects = ["../test/**/*"];
 
 export default defineNuxtConfig({
   compatibilityDate: "2026-08-30",
