+++ shared/sse.ts
